Refactor some code, cleanup , didnt remmember what I did

+++ Kristoslav-project/Assets/Scripts/DamageOnTouch.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class DamageOnTouch : MonoBehaviour
 {
    [SerializeField] 
-   float damage = 1;
+   int damage = 1;
     void OnCollisionEnter(Collision collisionInfo)
     {
         var character = collisionInfo.gameObject.GetComponentInChildren<Character>();
